Consolidate duplicated critical-curve plot helper into plot_utils
_compute_critical_curve_lines was copy-pasted verbatim into both
autolens/imaging/plot/fit_imaging_plots.py and
autolens/interferometer/plot/fit_interferometer_plots.py, and the two copies
had drifted: the imaging copy narrowly caught (ModuleNotFoundError, ValueError)
and loudly logged anything else (guarding the PyAutoGalaxy abd7b717 / PyAutoFit
#1280 silent-fallback regression), while the interferometer copy still had a
bare 'except Exception' that silently swallowed every failure — the exact mode
that regression guard exists to prevent. Model code imported inconsistent
behaviour: interferometer/model/{visualizer,plotter} got the silent copy,
point/model/* got the hardened one.

Move the single canonical (hardened) definition to autolens/plot/plot_utils.py
— which already re-exports the underlying _critical_curves_from/_caustics_from
primitives — and re-export it from both fit-plot modules so every existing
'from ...fit_{imaging,interferometer}_plots import _compute_critical_curve_lines'
import keeps resolving. All call sites now share one hardened definition.

Behaviour-preserving except on the interferometer error path, which is
intentionally lifted from silent-swallow to the hardened logging behaviour.
_plot_source_plane is deliberately left duplicated: the two copies genuinely
differ (mask source, zoom_extent_scale), so consolidating it is not
behaviour-neutral.

Witness tests (imaging error-path silent/logged assertions) updated to patch
and observe the symbol at its new home; interferometer plot tests do not pin
the error path, so none break. Dead imports (_caustics_from, numpy_lines)
trimmed from both modules.
